requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.20.0
safetensors>=0.4.0
omegaconf>=2.3.0
optimum[quanto]>=1.17.0
Pillow>=9.5.0
numpy>=1.24.0
